Moves custom fields to fields.py
