+++ chowda/views.py
@@ -1,4 +1,3 @@
-from dataclasses import dataclass
 from datetime import datetime, timedelta
 from typing import Any, ClassVar, Dict, List
 
@@ -6,54 +5,19 @@ from metaflow import Flow
 from metaflow.exception import MetaflowNotFound
 from metaflow.integrations import ArgoEvent
 from sqlmodel import Session, select
-from starlette.datastructures import FormData
 from starlette.requests import Request
 from starlette.responses import Response
 from starlette.templating import Jinja2Templates
 from starlette_admin import CustomView, action
-from starlette_admin._types import RequestAction
 from starlette_admin.contrib.sqlmodel import ModelView
 from starlette_admin.exceptions import ActionFailed
-from starlette_admin.fields import BaseField, IntegerField, TextAreaField
+from starlette_admin.fields import BaseField
 
 from chowda.auth.utils import get_user
 from chowda.db import engine
+from chowda.fields import MediaFileCount, MediaFilesGuidsField
 from chowda.models import Batch, Collection
 from chowda.utils import validate_media_file_guids
-
-
-@dataclass
-class MediaFilesGuidsField(TextAreaField):
-    """A field that displays a list of MediaFile GUIDs as html links"""
-
-    render_function_key: str = 'media_file_guid_links'
-    form_template: str = 'forms/media_file_guids_textarea.html'
-
-    async def parse_form_data(
-        self, request: Request, form_data: FormData, action: RequestAction
-    ) -> Any:
-        """Maps a string of GUID to a list"""
-        return form_data.get(self.id).split()
-
-    async def serialize_value(
-        self, request: Request, value: Any, action: RequestAction
-    ) -> Any:
-        """Maps a Collection's MediaFile objects to a list of GUIDs"""
-        return [media_file.guid for media_file in value]
-
-
-@dataclass
-class MediaFileCount(IntegerField):
-    """A field that displays the number of MediaFiles in a collection or batch"""
-
-    name: str = 'media_file_count'
-    label: str = 'Size'
-    read_only: bool = True
-    exclude_from_edit: bool = True
-    exclude_from_create: bool = True
-
-    async def parse_obj(self, request: Request, obj: Any) -> Any:
-        return len(obj.media_files)
 
 
 class BaseModelView(ModelView):
